+++ src/notes/accidental/utils/ops.rs
@@ -1,30 +1,30 @@
-use std::ops::{Add, Sub, AddAssign, SubAssign};
 use super::*;
+use std::ops::{Add, AddAssign, Sub, SubAssign};
 
-impl Add<i8> for Accidental {
+impl Add<i32> for Accidental {
     type Output = Self;
 
-    fn add(self, rhs: i8) -> Self::Output {
-        Self::from(i8::from(self) + rhs)
+    fn add(self, rhs: i32) -> Self::Output {
+        Self::from(i32::from(self) + rhs)
     }
 }
 
-impl AddAssign<i8> for Accidental {
-    fn add_assign(&mut self, rhs: i8) {
+impl AddAssign<i32> for Accidental {
+    fn add_assign(&mut self, rhs: i32) {
         *self = *self + rhs;
     }
 }
 
-impl Sub<i8> for Accidental {
+impl Sub<i32> for Accidental {
     type Output = Self;
-    
-    fn sub(self, rhs: i8) -> Self::Output {
-        Self::from(i8::from(self) - rhs)
+
+    fn sub(self, rhs: i32) -> Self::Output {
+        Self::from(i32::from(self) - rhs)
     }
 }
 
-impl SubAssign<i8> for Accidental {
-    fn sub_assign(&mut self, rhs: i8) {
+impl SubAssign<i32> for Accidental {
+    fn sub_assign(&mut self, rhs: i32) {
         *self = *self - rhs;
     }
 }
